backport of commit 0cc49cbd12cdd69bed62b9d846649be4443dfe85
